requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
cloudinary==1.44.2
Django==6.0.4
django-ckeditor==6.7.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-jazzmin==3.0.4
django-js-asset==3.1.2
djangorestframework==3.17.1
idna==3.13
pillow==12.2.0
psycopg2-binary==2.9.12
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
tzdata==2026.2
urllib3==2.6.3
